backend/pyproject.toml
[project]
name = "IspBirntg-backend"
version = "0.1.0"
description = "離線版 ChatPDF 後端服務"
requires-python = ">= 3.9"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.0",
    "llama-index>=0.9",
    "llama-index-embeddings-huggingface>=0.1",
    "llama-index-llms-gemini>=0.1",
    "PyMuPDF>=1.23",
    "python-dotenv>=1.0",
    "google-generativeai>=0.3",
    "cryptography>=41.0",
    "sentence-transformers>=2.2",
    "torch>=2.0",
    "pillow>=10.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/IspBirntg"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "black>=23.0",
    "flake8>=6.0",
]
